models/OpenCLIP/requirements.txt
numpy
transformers
pillow